Add a read-only mode for houston.

+++ src/martian/houston/submissions.go
@@ -70,11 +70,12 @@ type SubmissionManager struct {
 	cache                  map[string]*Submission
 	rt                     *core.Runtime
 	mailer                 *manager.Mailer
+	readonly               bool
 }
 
 func NewSubmissionManager(hostname string, instanceName string, filesPath string,
 	cachePath string, pipestanceSummaryPaths []string, rt *core.Runtime,
-	mailer *manager.Mailer) *SubmissionManager {
+	mailer *manager.Mailer, readonly bool) *SubmissionManager {
 	self := &SubmissionManager{}
 	self.instanceName = instanceName
 	self.filesPath = filesPath
@@ -85,6 +86,7 @@ func NewSubmissionManager(hostname string, instanceName string, filesPath string
 	self.mailer = mailer
 	self.loadCache()
 	self.hostname = hostname
+	self.readonly = readonly
 	return self
 }
 
@@ -257,8 +259,10 @@ func (self *SubmissionManager) InventorySubmissions() {
 		self.mailer.Sendmail(users, subj, body)
 	}
 
-	// Write submissions to persistent cache
-	writeJson(self.cachePath, self.cache)
+	if !self.readonly {
+		// Write submissions to persistent cache
+		writeJson(self.cachePath, self.cache)
+	}
 }
 
 func (self *SubmissionManager) EnumerateSubmissions() []*Submission {
